% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtx.R
\name{dtx}
\alias{dtx}
\alias{dtrm}
\alias{dtbl}
\alias{dtf}
\alias{dtt}
\title{Create a data frame (data.trame, base's data.frame, data.table or tibble's tbl_df)}
\usage{
dtx(..., .name_repair = c("check_unique", "unique", "universal", "minimal"))

dtrm(..., .name_repair = c("check_unique", "unique", "universal", "minimal"))

dtbl(..., .name_repair = c("check_unique", "unique", "universal", "minimal"))

dtf(..., .name_repair = c("check_unique", "unique", "universal", "minimal"))

dtt(..., .name_repair = c("check_unique", "unique", "universal", "minimal"))
}
\arguments{
\item{...}{A set of name-value pairs. The content of the data frame. See
\code{\link[tibble:tibble]{tibble::tibble()}} for more details on the way dynamic-dots are processed.}

\item{.name_repair}{The way problematic column names are treated, see also
\code{\link[tibble:tibble]{tibble::tibble()}} for details.}
}
\value{
A data frame as a \strong{data.trame} object for \code{\link[=dtrm]{dtrm()}}, a \strong{tbl_df} object
for \code{\link[=dtbl]{dtbl()}}, a \strong{data.frame} for \code{\link[=dtf]{dtf()}} or a \strong{data.table} for \code{\link[=dtt]{dtt()}}.
}
\description{
Create a data frame (data.trame, base's data.frame, data.table or tibble's tbl_df)
}
\note{
data.trame, data.table and tibble's tbl_df do no use row names.
However, you can add a column named \code{.rownames}(by default), or the name that
is in \code{getOption("SciViews.dtx.rownames")} and it will be automatically set as
row names when the object is converted into a data.frame with \code{\link[=as_dtf]{as_dtf()}}. For
\code{\link[=dtf]{dtf()}}, just create a column of this name and it is directly used as row
names for the resulting data.frame object.
}
\examples{
dtrm1 <- dtrm(
  x = 1:5,
  y = rnorm(5),
  f = letters[1:5],
  l = sample(c(TRUE, FALSE), 5, replace = TRUE)
)
class(dtrm1)

dtbl1 <- dtbl(
  x = 1:5,
  y = rnorm(5),
  f = letters[1:5],
  l = sample(c(TRUE, FALSE), 5, replace = TRUE)
)
class(dtbl1)

dtf1 <- dtf(
  x = 1:5,
  y = rnorm(5),
  f = letters[1:5],
  l = sample(c(TRUE, FALSE), 5, replace = TRUE)
)
class(dtf1)

dtt1 <- dtt(
  x = 1:5,
  y = rnorm(5),
  f = letters[1:5],
  l = sample(c(TRUE, FALSE), 5, replace = TRUE))
class(dtt1)

# Using dtx(), one construct the preferred data frame object
# (a data.trame by default, can be changed with options(SciViews.as_dtx = ...))
dtx1 <- dtx(
  x = 1:5,
  y = rnorm(5),
  f = letters[1:5],
  l = sample(c(TRUE, FALSE), 5, replace = TRUE))
class(dtx1) # data.trame by default

# Use dtx_rows() to easily create a data frame:
dtx2 <- dtx_rows(
  ~x, ~y, ~f,
   1,  3, 'a',
   2,  4, 'b'
)
dtx2
class(dtx2)

# This is how you specify row names for dtf (data.frame)
dtf(x = 1:3, y = 4:6, .rownames = letters[1:3])
}
\seealso{
\code{\link[=dtx_rows]{dtx_rows()}}, \code{\link[=is_dtx]{is_dtx()}}, \code{\link[=collect_dtx]{collect_dtx()}}
}
