% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtx_rows.R
\name{dtx_rows}
\alias{dtx_rows}
\alias{dtf_rows}
\alias{dtt_rows}
\alias{dtbl_rows}
\title{Row-wise creation of a data frame}
\usage{
dtx_rows(...)

dtf_rows(...)

dtt_rows(...)

dtbl_rows(...)
}
\arguments{
\item{...}{Specify the structure of the data frame by using formulas for
variable names like \code{~x} for variable \code{x}. Then, use one argument per value
in the data frame. It is possible to unquote with \verb{!!} and to
unquote-splice with \verb{!!!}.}
}
\value{
A data frame of class \strong{data.frame} for \code{\link[=dtf_rows]{dtf_rows()}}, \strong{data.table}
for \code{\link[=dtt_rows]{dtt_rows()}}, tibble \strong{tbl_df} for \code{\link[=dtbl_rows]{dtbl_rows()}} and the default object
with \code{\link[=dtx_rows]{dtx_rows()}}.
}
\description{
The presentation of the data (see examples) is easier to read
than with the traditional column-wise entry in \code{\link[=dtx]{dtx()}}. This could be used
to enter small tables in R, but do not abuse of it!
}
\examples{
df <- dtx_rows(
  ~x, ~y, ~group,
   1,  3,    "A",
   6,  2,    "A",
   10, 4,    "B"
)
df
}
