% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_.R
\name{f_}
\alias{f_}
\title{Convert a string into a formula and/or change the associated environment}
\usage{
f_(x, env = parent.frame())
}
\arguments{
\item{x}{Either a name (character string) or a formula}

\item{env}{The environment to associate with the formula}
}
\value{
A formula
}
\description{
This function is used to transform a name like \code{'var'} into a formula \code{~var}.
}
\examples{
f_('var')
}
