% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_functions.R
\name{fast_functions}
\alias{fast_functions}
\alias{fn}
\alias{fnobs_all}
\alias{fnobs_na}
\alias{\%replacef\%}
\alias{\%replace_fillf\%}
\alias{\%-f\%}
\alias{\%+f\%}
\alias{\%-+f\%}
\alias{\%/f\%}
\alias{\%/*100f\%}
\alias{\%*f\%}
\alias{\%modf\%}
\alias{\%-modf\%}
\title{Fast (flexible and friendly) statistical functions (mainly from {collapse}) for matrix-like and data frame objects}
\usage{
fast_functions()

fn(x, ...)

fnobs_all(x, ...)

fnobs_na(x, ...)

x \%replacef\% expr

x \%replace_fillf\% expr

x \%-f\% expr

x \%+f\% expr

x \%-+f\% expr

x \%/f\% expr

x \%/*100f\% expr

x \%*f\% expr

x \%modf\% expr

x \%-modf\% expr
}
\arguments{
\item{x}{A numeric vector, matrix, data frame or grouped data frame (class
'grouped_df').}

\item{...}{Further arguments passed to the method, like \verb{w=}, a numeric
vector of (non-negative) weights that may contain missing values, or
\verb{TRA=}, a quoted operator indicating the transformation to perform:
\code{"replace"} to get a vector of same size of \code{x} with results,
\code{"replace_fill"} idem but also replace missing data, \code{"-"} to subtract,
\code{"+"} to add, \code{"-+"} to subtract and add the global statistic, \code{"/"}
to divide, \code{"\%"} to divide and multiply by 100 (percent), \code{"*"} to
multiply, \code{"\%\%"} to take the modulus (remainder from division by the
statistic) and \code{"-\%\%"} to subtract modulus ('i.e., to floor the data by the
statistic), see \code{\link[collapse:TRA]{collapse::TRA()}}. Also \verb{na.rm=}, a logical indicating if
we skip missing values in \code{x}if \code{TRUE}(by default). If \code{FALSE} for any
missing data in \code{x}, \code{NA}is returned. For details and other arguments,
see the corresponding help page in, the {collapse} package.}

\item{expr}{The expression to evaluate as RHS of the \verb{\%__f\%} operators.}
}
\value{
The number of all observations for \code{\link[=fnobs_all]{fnobs_all()}} or the number of
missing observations for \code{\link[=fnobs_na]{fnobs_na()}}. \code{\link[=fast_functions]{fast_functions()}} returns a list of
all the known fast statistical functions.
}
\description{
The fast statistical function, or fast-flexible-friendly
statistical functions are prefixed with "f". These vectorized functions
supersede the no-f functions, bringing the capacity to work smoothly on
matrix-like and data frame objects. For instance, base \code{\link[=mean]{mean()}} operates on a
vector, but not on a data frame. A matrix is recognized as a vector and a
single mean is returned. On, the contrary, \code{\link[=fmean]{fmean()}} calculates one mean per
column. It does the same for a data frame, and it does so usually quicker
than base functions. No need for \code{colMeans()}, a separate function to do so.
Fast functions also recognize grouping with \code{\link[=fgroup_by]{fgroup_by()}}, \code{\link[=sgroup_by]{sgroup_by()}},
\code{\link[=tgroup_by]{tgroup_by()}} or \code{\link[=group_by]{group_by()}} and calculate the mean by group in this case.
Again, no need for a different function like \code{\link[stats:ave]{stats::ave()}}.
Finally, these functions also have a \verb{TRA=} argument that computes, for
instance, if \code{TRA = "-"}, \verb{(x  f(x))} very efficiently (for instance to
calculate residuals by subtracting the mean).
Another particularity is the \verb{na.rm=} argument that is \code{TRUE} by default,
while it is \code{FALSE} by default for \code{\link[=mean]{mean()}}.
These are generic functions with methods for \strong{matrix}, \strong{data.frame},
\strong{grouped_df} and a \strong{default} method used for simple numeric vectors. Most
of them are defined in the {collapse} package, but there are a couple more
here, together with an alternate syntax to replace \verb{TRA=} with \verb{\%_f\%}.
}
\note{
The page \link[collapse:fast-statistical-functions]{collapse::fast-statistical-functions} gives more details.
\code{\link[=fnobs_all]{fnobs_all()}} count all observations, including \code{NA}s, \code{\link[=fnobs_na]{fnobs_na()}} counts
only \code{NA}s, where \code{\link[=fnobs]{fnobs()}} counts non-missing observations.
Instead of \verb{TRA=} one can use the \verb{\%__f\%} functions where \verb{__} is \code{replace}, \code{replace_fill}, \code{-}, \code{+}, \verb{-+}, \code{/}, \verb{/*100} for \code{TRA="\%"}, \code{*}, \code{mod} for \code{TRA="\%\%"}, or \code{-mod} for \code{TRA="-\%\%"}. See example.
}
\examples{
library(collapse)
data(iris)
iris_num <- iris[, -5] # Only numerical variables
mean(iris$Sepal.Length) # OK, but mean(iris_num does not work)
colMeans(iris_num)
# Same
fmean(iris_num)
# Idem, but mean by group for all 4 numerical variables
iris |> fgroup_by(Species) |> fmean()
# Residuals (x - mean(x)) by group
iris |> fgroup_by(Species) |> fmean(TRA = "-")
# The same calculation, in a little bit more expressive way
iris |> fgroup_by(Species) \%-f\% fmean()
# or:
iris_num \%-f\% fmean(g = iris$Species)
}
