% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_select.R
\name{formula_select}
\alias{formula_select}
\title{Formula-select is similar to tidy-select in the tidyverse, but using formulas}
\usage{
formula_select(
  ...,
  .fast.allowed.funs = NULL,
  .max.args = NULL,
  .must.be.named = FALSE,
  .make.names = FALSE,
  .no.se = FALSE,
  .no.se.msg = gettext("Standard evaluation is not allowed."),
  .envir = parent.frame(2L),
  .frame = parent.frame()
)
}
\arguments{
\item{...}{Arguments to be processed by formula-masking.}

\item{.fast.allowed.funs}{A character vector of function names that are
allowed for a fast treatment (usually though collapse functions). If any
other function is used, the slower tidy-select mechanism is used, see
\code{\link[tidyselect:eval_select]{tidyselect::eval_select()}}.}

\item{.max.args}{The maximum allowed arguments in \code{...}.}

\item{.must.be.named}{If \code{TRUE}, all arguments must be named.}

\item{.make.names}{If \code{TRUE}, unnamed arguments are named automatically.}

\item{.no.se}{If \code{TRUE}, standard evaluation is not allowed.}

\item{.no.se.msg}{The message to be used if standard evaluation is not
allowed.}

\item{.envir}{The environment where to expand formulas (possibly superseded
by the environment attached to the first formula).}

\item{.frame}{The frame where the focus in the calling stack should be set
in error messages (not used yet).}
}
\value{
A list with components:
\itemize{
\item \code{dots}: the processed arguments (formulas are turned into expressions)
\item \code{are_formulas}: whether the arguments were formulas
\item \code{env}: The environment where the expressions should be evaluated.
\item \code{fastselect}: whether fast selection can be used
}
}
\description{
The formula-select interface allows to give arguments in the tidy-select
style withing formulas, or as standard-evaluated arguments. It thus combines
both approaches within the same function, and makes it clearer which are the
intention: standard evaluation, or non-standard evaluation when formulas are
used.
}
\examples{
# TODO...
}
\seealso{
\code{\link[=formula_masking]{formula_masking()}},
\url{https://tidyselect.r-lib.org/articles/syntax.html} or
\code{vignette("syntax", package = "tidyselect")} for a technical
description of the rules of evaluation.
}
