% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_dtx.R
\name{is_dtx}
\alias{is_dtx}
\alias{is_dtf}
\alias{is_dtt}
\alias{is_dtbl}
\title{Test if the object is a data frame (data.frame, data.table or tibble)}
\usage{
is_dtx(x, strict = TRUE)

is_dtf(x, strict = TRUE)

is_dtt(x, strict = TRUE)

is_dtbl(x, strict = TRUE)
}
\arguments{
\item{x}{An object}

\item{strict}{Should this be strictly the corresponding class \code{TRUE}, by default, or could it be subclassed too (\code{FALSE}). With \code{strict = TRUE}, the \strong{grouped_df} tibbles and \strong{grouped_ts} tsibbles are also considered (tibbles or tsibbles where \code{\link[dplyr:group_by]{dplyr::group_by()}} was applied).}
}
\value{
These functions return \code{TRUE} if the object is of the correct class, otherwise they return \code{FALSE}. \code{is_dtx()} return \code{TRUE} if \code{x} is one of a data.frame, data.table or tibble.
}
\description{
Test if the object is a data frame (data.frame, data.table or tibble)
}
\examples{
# data(mtcars)
is_dtf(mtcars) # TRUE
is_dtx(mtcars) # Also TRUE
is_dtt(mtcars) # FALSE
is_dtbl(mtcars) # FALSE
# but...
is_dtt(as_dtt(mtcars)) # TRUE
is_dtx(as_dtt(mtcars)) # TRUE
is_dtbl(as_dtbl(mtcars)) # TRUE
is_dtx(as_dtbl(mtcars)) # TRUE
is_dtx(as_dtbl(mtcars) |> dplyr::group_by(cyl)) # TRUE (special case)

is_dtx("some string") # FALSE
}
