% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_error_msg.R
\name{process_error_msg}
\alias{process_error_msg}
\title{Process error message (replace it with a better version)}
\usage{
process_error_msg(msg, tra, fixed = TRUE)
}
\arguments{
\item{msg}{The original error message.}

\item{tra}{The translation list, where the names are the partial error and
the content is the new error message(s).}

\item{fixed}{Should a fixed or a regular expression comparison be done
between the names of \code{tra} and the \code{msg}? Defaults to \code{TRUE}, which is
faster. In case you need different behavior for different items, you can
also supply a logical vector of the same length as \code{tra}.}
}
\value{
The replaced error message.
}
\description{
The \code{tra} list indicates which (partial) error message should be replaced
by a more elaborate version, using the \code{\link[cli:cli_abort]{cli::cli_abort()}} syntax.
}
\examples{
dir <- "/temp/dir"
file <- "~/some file.txt"
tra = list(
 "Directory not found" = c("The directory {.path {dir}} does not exist.",
   i = "Please create it first."),
 "file does not exist" = c("The file {.file {file}} does not exist.",
   i = "Check the path and try again.")
)
msg <- "The file does not exist"
try(cli::cli_abort(process_error_msg(msg, tra)))

msg <- "Directory not found on this machine"
try(cli::cli_abort(process_error_msg(msg, tra)))
}
