% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedy_functions.R
\name{speedy_functions}
\alias{speedy_functions}
\alias{sgroup_by}
\alias{sungroup}
\alias{srename}
\alias{srename_with}
\alias{sfilter}
\alias{sfilter_ungroup}
\alias{sselect}
\alias{smutate}
\alias{smutate_ungroup}
\alias{stransmute}
\alias{stransmute_ungroup}
\alias{ssummarise}
\alias{sfull_join}
\alias{sleft_join}
\alias{sright_join}
\alias{sinner_join}
\title{Speedy functions (mainly from {collapse} and {data.table}) to manipulate data frames}
\usage{
speedy_functions()

sgroup_by(.data, ...)

sungroup(.data, ...)

srename(.data, ...)

srename_with(.data, .fn, .cols = everything(), ...)

sfilter(.data, ...)

sfilter_ungroup(.data, ...)

sselect(.data, ...)

smutate(.data, ..., .keep = "all")

smutate_ungroup(.data, ..., .keep = "all")

stransmute(.data, ...)

stransmute_ungroup(.data, ...)

ssummarise(.data, ...)

sfull_join(x, y, by = NULL, suffix = c(".x", ".y"), copy = FALSE, ...)

sleft_join(x, y, by = NULL, suffix = c(".x", ".y"), copy = FALSE, ...)

sright_join(x, y, by = NULL, suffix = c(".x", ".y"), copy = FALSE, ...)

sinner_join(x, y, by = NULL, suffix = c(".x", ".y"), copy = FALSE, ...)
}
\arguments{
\item{.data}{A data frame (data.frame, data.table or tibble's tbl_df)}

\item{...}{Arguments dependent to the context of the function and most of
the time, not evaluated in a standard way (cf. the tidyverse approach).}

\item{.fn}{A function to use.}

\item{.cols}{The list of the column where to apply the transformation. For
the moment, only all existing columns, which means \code{.cols = everything()}
is implemented}

\item{.keep}{Which columns to keep. The default is \code{"all"}, possible values
are \code{"used"}, \code{"unused"}, or \code{"none"} (see \code{\link[=mutate]{mutate()}}).}

\item{x}{A data frame (data.frame, data.table or tibble's tbl_df).}

\item{y}{A second data frame.}

\item{by}{A list of names of the columns to use for joining the two data
frames.}

\item{suffix}{The suffix to the column names to use to differentiate the
columns that come from the first or the second data frame. By default it is
\code{c(".x", ".y")}.}

\item{copy}{This argument is there for compatibility with the "t" matching
functions, but it is not used here.}
}
\value{
See corresponding "non-s" function for the full help page with
indication of the return values.
}
\description{
The Tidyverse defines a coherent set of tools to manipulate
data frames that use a non-standard evaluation and sometimes require extra
care. These functions, like \code{\link[=mutate]{mutate()}} or \code{\link[=summarise]{summarise()}} are defined in the
{dplyr} and {tidyr} packages. The {collapse} package proposes a couple of
functions with similar interface, but with different and much faster code.
For instance, \code{\link[=fselect]{fselect()}} is similar to \code{\link[=select]{select()}}, or \code{\link[=fsummarise]{fsummarise()}} is
similar to \code{\link[=summarise]{summarise()}}. Not all functions are implemented, arguments and
argument names differ, and the behavior may be very different, like
\code{\link[=frename]{frename()}} which uses \code{old_name = new_name}, while \code{\link[=rename]{rename()}} uses
\code{new_name = old_name}! The speedy functions all start with an "s" prefixed to
the function name, like \code{\link[=smutate]{smutate()}}, and build on the work initiated in
{collapse} to propose a series of paired functions with the tidy ones. So,
\code{\link[=smutate]{smutate()}} and \code{\link[=tmutate]{tmutate()}} are "speedy" and 'tidy" counterparts and they are
used in a very similar, if not identical way as the original \code{\link[=mutate]{mutate()}}
function. This notation using a "s" prefix is there to draw the attention on
their particularities. Their classes are \strong{function} and \strong{speedy_fn}.
Avoid mixing tidy, speedy and non-tidy/speedy functions in the same pipeline.
}
\note{
The help page here is very basic and it aims mainly to list all the
speedy functions. For more complete help, see their "non-s" counterparts in
{dplyr} or {tidyr}, or use the {svMisc}'s \code{.?smutate} syntax to link to the
correct page.
}
\examples{
# TODO...
}
