% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svBase-package.R
\docType{package}
\name{svBase-package}
\alias{svBase}
\alias{svBase-package}
\title{svBase: 'SciViews' - Functions for Base Objects like Data Frames}
\description{
Functions to manipulated the three main classes of "data frames" for 'SciViews': data.frame, data.table and tibble. Allow to select the preferred one, and to convert more carefully between the three, taking care of correct presentation of row names and data.table's keys. More homogeneous way of creating these three data frames and of printing them on the R console.
}
\details{
The {svBase} package sets up the way data frames (with objects like R base's \strong{data.frame}, \strong{data.table} and tibble \strong{tbl_df}) are managed in SciViews. The user can select the class of object it uses by default and many other SciViews functions return that format. Also conversion from one to the other is smoothed, including for the management of \strong{data.frame}'s row names or \strong{data.table}'s keys. Also homogeneous ways to create a data frame or to print it are also provided.
}
\section{Important functions}{

\itemize{
\item \code{\link[=dtx]{dtx()}} creates a data frame in the preferred format, with the
following functions \code{\link[=dtbl]{dtbl()}}, \code{\link[=dtf]{dtf()}} and \code{\link[=dtt]{dtt()}} that force respectively
the creation of a data frame in one of the specified three formats. Use
\code{getOption("SciViews.as_dtx", default = as_dtt)} to specify which function to
use to convert into the preferred format.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SciViews/svBase}
  \item \url{https://www.sciviews.org/svBase/}
  \item Report bugs at \url{https://github.com/SciViews/svBase/issues}
}

}
\author{
\strong{Maintainer}: Philippe Grosjean \email{phgrosjean@sciviews.org} (\href{https://orcid.org/0000-0002-2694-9471}{ORCID})

}
\keyword{internal}
