% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtx.R
\name{dtx}
\alias{dtx}
\alias{dtbl}
\alias{dtf}
\alias{dtt}
\title{Create a data frame (base's data.frame, data.table or tibble's tbl_df)}
\usage{
dtx(..., .name_repair = c("check_unique", "unique", "universal", "minimal"))

dtbl(..., .name_repair = c("check_unique", "unique", "universal", "minimal"))

dtf(..., .name_repair = c("check_unique", "unique", "universal", "minimal"))

dtt(..., .name_repair = c("check_unique", "unique", "universal", "minimal"))
}
\arguments{
\item{...}{A set of name-value pairs. The content of the data frame. See
\code{\link[=tibble]{tibble()}} for more details on the way dynamic-dots are processed.}

\item{.name_repair}{The way problematic column names are treated, see also
\code{\link[=tibble]{tibble()}} for details.}
}
\value{
A data frame as a \strong{tbl_df} object for \code{\link[=dtbl]{dtbl()}}, a \strong{data.frame} for
\code{\link[=dtf]{dtf()}} and a \strong{data.table} for \code{\link[=dtt]{dtt()}}.
}
\description{
Create a data frame (base's data.frame, data.table or tibble's tbl_df)
}
\examples{
dtbl1 <- dtbl(
  x = 1:5,
  y = rnorm(5),
  f = letters[1:5],
  l = sample(c(TRUE, FALSE), 5, replace = TRUE)
)
class(dtbl1)

dtf1 <- dtf(
  x = 1:5,
  y = rnorm(5),
  f = letters[1:5],
  l = sample(c(TRUE, FALSE), 5, replace = TRUE)
)
class(dtf1)

dtt1 <- dtt(
  x = 1:5,
  y = rnorm(5),
  f = letters[1:5],
  l = sample(c(TRUE, FALSE), 5, replace = TRUE))
class(dtt1)
}
