% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtx.R
\name{dtx}
\alias{dtx}
\alias{dtbl}
\alias{dtf}
\alias{dtt}
\alias{show.data.frame}
\alias{tbl_sum.dataframe}
\alias{show.data.table}
\alias{tbl_sum.datatable}
\title{Create a data frame (base's data.frame, data.table or tibble's tbl_df)}
\usage{
dtx(..., .name_repair = c("check_unique", "unique", "universal", "minimal"))

dtbl(..., .name_repair = c("check_unique", "unique", "universal", "minimal"))

dtf(..., .name_repair = c("check_unique", "unique", "universal", "minimal"))

dtt(..., .name_repair = c("check_unique", "unique", "universal", "minimal"))

\method{show}{data.frame}(x, ...)

\method{tbl_sum}{dataframe}(x, ...)

\method{show}{data.table}(x, ...)

\method{tbl_sum}{datatable}(x, ...)
}
\arguments{
\item{...}{A set of name-value pairs. The content of the data frame. See
\code{\link[=tibble]{tibble()}} for more details on the way dynamic-dots are processed.}

\item{.name_repair}{The way problematic column names are treated, see also
\code{\link[=tibble]{tibble()}} for details.}

\item{x}{An object to print.}
}
\value{
A data frame as a \strong{tbl_df} object for \code{\link[=dtbl]{dtbl()}}, a \strong{data.frame} for
\code{\link[=dtf]{dtf()}} and a \strong{data.table} for \code{\link[=dtt]{dtt()}}.
}
\description{
Create a data frame (base's data.frame, data.table or tibble's tbl_df)
}
\note{
data.table and tibble's tbl_df do no use row names. However, you can
add a column named \code{.rownames}(by default), or the name that is in
\code{getOption("SciViews.dtx.rownames")} and it will be automatically set as row
names when the object is converted into a data.frame with \code{\link[=as_dtf]{as_dtf()}}. For
\code{\link[=dtf]{dtf()}}, just create a column of this name and it is directly used as row
names for the resulting data.frame object.
}
\examples{
dtbl1 <- dtbl(
  x = 1:5,
  y = rnorm(5),
  f = letters[1:5],
  l = sample(c(TRUE, FALSE), 5, replace = TRUE)
)
class(dtbl1)

dtf1 <- dtf(
  x = 1:5,
  y = rnorm(5),
  f = letters[1:5],
  l = sample(c(TRUE, FALSE), 5, replace = TRUE)
)
class(dtf1)

dtt1 <- dtt(
  x = 1:5,
  y = rnorm(5),
  f = letters[1:5],
  l = sample(c(TRUE, FALSE), 5, replace = TRUE))
class(dtt1)

# Using dtx(), one construct the preferred data frame object
# (a data.table by default, can be changed with options(SciViews.as_dtx = ...))
dtx1 <- dtx(
  x = 1:5,
  y = rnorm(5),
  f = letters[1:5],
  l = sample(c(TRUE, FALSE), 5, replace = TRUE))
class(dtx1) # data.table by default

# With {svBase} data.table and data.frame objects have the same nice print as tibbles
dtbl1
dtf1
dtt1

# Use tribble() inside dtx() to easily create a data frame:
library(tibble)
dtx2 <- dtx(tribble(
  ~x, ~y, ~f,
   1,  3, 'a',
   2,  4, 'b'
))
dtx2
class(dtx2)

# This is how you specify row names for dtf (data.frame)
dtf(x = 1:3, y = 4:6, .rownames = letters[1:3])
}
